Delete confirmation dialog

+++ EmployeeManagement.Web/Pages/EditEmployeeBase.cs
@@ -113,10 +113,26 @@ namespace EmployeeManagement.Web.Pages
             }
         }
 
-        protected async Task Delete_Click()
+        protected BlazorProject.Components.ConfirmBase DeleteConfirmation { get; set; }
+
+        protected void Delete_Click()
         {
-            await EmployeeService.DeleteEmployee(Employee.EmployeeId);
-            NavigationManager.NavigateTo("/");
+            DeleteConfirmation.Show();
         }
+
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if (deleteConfirmed)
+            {
+                await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+                NavigationManager.NavigateTo("/");
+            }
+        }
+
+        //protected async Task Delete_Click()
+        //{
+        //    await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+        //    NavigationManager.NavigateTo("/");
+        //}
     }
 }
